Add files via upload
This was my first project for this Nanodegree. The project was about analysing data about the medical appointment bookings of patients in Brazil. There are fourteen columns in the dataset, thirteen predictor variables and one target variable - "No Show". The target variable indicates if the patient attended his/her medical appointment or not. 

Based on the information available in the dataset, we were required to consider the possible questions that could be asked about the factors that influence a patient's ability to honour medical appointments.