Add files via upload
This was the second project for the Nanodegree.  For this project, I had to gather data from three different sources and merge them into one dataset because I could carry out the required analysis. The first set of data was made available as a "csv" file. The second set of data had to be downloaded programmatically from the Udacity server using the request library. The third set of data had to be scrapped from the Twitter account of the Twitter user "WeRateDogs". 

After gathering all three pieces of data, they were wrangled and merged, leaving behind only the columns required for the analysis.